+++ core/class/AbeilleParser.class.php
@@ -25,7 +25,7 @@
             global $argv;
 
             /* Configuring log library to use 'logMessage()' */
-            logSetConf("AbeilleParser.log", true);
+            // logSetConf("AbeilleParserD.log", true); // Tcharp38: Required ? Already done by AbeilleParserD
 
             parserLog("debug", "AbeilleParser constructor", "AbeilleParserClass");
             // $this->parameters_info = AbeilleTools::getParameters();