+++ docs/fr_FR/Changelog.md
@@ -6,6 +6,7 @@
 -   Améliorations: Page config. Amélioration mineure pour lien Github.
 -   Améliorations: Ajout LQI dans trace debug parser.
 -   Améliorations: Page config. Liste des ports revue. Suppression utilisation getUsbMapping() bugguée.
+-   Améliorations: Interne, mineurg: AbeilleCmd/Parser => AbeilleCmdD/ParserD.
 
 ## 251024-BETA-1